Create 2_3
Write a program that creates an 8x8 grid. Use the new line to separate lines. Each position should be " " or "#". Everything should form a style that resembles the pattern of a chess board. You should be able to change the value of the size variable to change the output of the width and height of the grid.